requirements.txt
pytest
pytest-cov
pytest-pylint